+++ src/Projects/index.ts
@@ -1,6 +1,12 @@
 import ApiGroup, { ApiConfig } from '../ApiGroup';
-import models from './models.json';
-import { AvailableModel, EstimateRequest, ImageUrlParams, ProjectParams } from './types';
+import {
+  AvailableModel,
+  EstimateRequest,
+  ImageUrlParams,
+  ProjectParams,
+  SizePreset,
+  SupportedModel
+} from './types';
 import {
   JobErrorData,
   JobProgressData,
@@ -16,8 +22,12 @@ import { JobEvent, ProjectApiEvents, ProjectEvent } from './types/events';
 import getUUID from '../lib/getUUID';
 import { RawProject } from './types/RawProject';
 import ErrorData from '../types/ErrorData';
+import { SupernetType } from '../ApiClient/WebSocketClient/types';
+import Cache from '../lib/Cache';
 
+const sizePresetCache = new Cache<SizePreset[]>(10 * 60 * 1000);
 const GARBAGE_COLLECT_TIMEOUT = 10000;
+const MODELS_REFRESH_INTERVAL = 1000 * 60 * 60 * 24; // 24 hours
 
 function mapErrorCodes(code: string): number {
   switch (code) {
@@ -39,6 +49,10 @@ function mapErrorCodes(code: string): number {
 class ProjectsApi extends ApiGroup<ProjectApiEvents> {
   private _availableModels: AvailableModel[] = [];
   private projects: Project[] = [];
+  private _supportedModels: { data: SupportedModel[] | null; updatedAt: Date } = {
+    data: null,
+    updatedAt: new Date(0)
+  };
 
   get availableModels() {
     return this._availableModels;
@@ -65,14 +79,20 @@ class ProjectsApi extends ApiGroup<ProjectApiEvents> {
     this.emit('availableModels', this._availableModels);
   }
 
-  private handleSwarmModels(data: SocketEventMap['swarmModels']) {
-    const modelIndex = models.reduce((acc: Record<string, any>, model) => {
-      acc[model.modelId] = model;
+  private async handleSwarmModels(data: SocketEventMap['swarmModels']) {
+    let models: SupportedModel[] = [];
+    try {
+      models = await this.getSupportedModels();
+    } catch (e) {
+      this.client.logger.error(e);
+    }
+    const modelIndex = models.reduce((acc: Record<string, SupportedModel>, model) => {
+      acc[model.id] = model;
       return acc;
     }, {});
     this._availableModels = Object.entries(data).map(([id, workerCount]) => ({
       id,
-      name: modelIndex[id]?.modelShortName || id.replace(/-/g, ' '),
+      name: modelIndex[id]?.name || id.replace(/-/g, ' '),
       workerCount
     }));
     this.emit('availableModels', this._availableModels);
@@ -339,6 +359,38 @@ class ProjectsApi extends ApiGroup<ProjectApiEvents> {
     return data.project;
   }
 
+  /**
+   * Cancel project by id. This will cancel all jobs in the project and mark project as canceled.
+   * Client may still receive job events for the canceled jobs as it takes some time, but they will
+   * be ignored
+   * @param projectId
+   **/
+  async cancel(projectId: string) {
+    await this.client.socket.send('jobError', {
+      jobID: projectId,
+      error: 'artistCanceled',
+      error_message: 'artistCanceled',
+      isFromWorker: false
+    });
+    const project = this.projects.find((p) => p.id === projectId);
+    if (!project) {
+      return;
+    }
+    // Remove project from the list to stop tracking it
+    this.projects = this.projects.filter((p) => p.id !== projectId);
+
+    // Cancel all jobs in the project
+    project.jobs.forEach((job) => {
+      if (!job.finished) {
+        job._update({ status: 'canceled' });
+      }
+    });
+    // If project is still in processing, mark it as canceled
+    if (!project.finished) {
+      project._update({ status: 'canceled' });
+    }
+  }
+
   private async uploadGuideImage(projectId: string, file: File | Buffer | Blob) {
     const imageId = getUUID();
     const presignedUrl = await this.uploadUrl({
@@ -370,10 +422,32 @@ class ProjectsApi extends ApiGroup<ProjectApiEvents> {
     stepCount,
     previewCount,
     cnEnabled,
-    startingImageStrength
+    startingImageStrength,
+    width,
+    height,
+    sizePreset
   }: EstimateRequest) {
+    const pathParams = [
+      network,
+      model,
+      imageCount,
+      stepCount,
+      previewCount,
+      cnEnabled ? 1 : 0,
+      startingImageStrength ? 1 - startingImageStrength : 0
+    ];
+    if (sizePreset) {
+      const presets = await this.getSizePresets(network, model);
+      const preset = presets.find((p) => p.id === sizePreset);
+      if (!preset) {
+        throw new Error('Invalid size preset');
+      }
+      pathParams.push(preset.width, preset.height);
+    } else if (width && height) {
+      pathParams.push(width, height);
+    }
     const r = await this.client.socket.get<EstimationResponse>(
-      `/api/v1/job/estimate/${network}/${model}/${imageCount}/${stepCount}/${previewCount}/${cnEnabled ? 1 : 0}/${startingImageStrength ? 1 - startingImageStrength : 0}`
+      `/api/v1/job/estimate/${pathParams.join('/')}`
     );
     return {
       token: r.quote.project.costInToken,
@@ -405,6 +479,74 @@ class ProjectsApi extends ApiGroup<ProjectApiEvents> {
       params
     );
     return r.data.downloadUrl;
+  }
+
+  async getSupportedModels(forceRefresh = false) {
+    if (
+      this._supportedModels.data &&
+      !forceRefresh &&
+      Date.now() - this._supportedModels.updatedAt.getTime() < MODELS_REFRESH_INTERVAL
+    ) {
+      return this._supportedModels.data;
+    }
+    const models = await this.client.socket.get<SupportedModel[]>(`/api/v1/models/list`);
+    this._supportedModels = { data: models, updatedAt: new Date() };
+    return models;
+  }
+
+  /**
+   * Get supported size presets for the model and network. Size presets are cached for 10 minutes.
+   *
+   * @example
+   * ```ts
+   * const presets = await client.projects.getSizePresets('fast', 'flux1-schnell-fp8');
+   * console.log(presets);
+   * ```
+   *
+   * @param network - 'fast' or 'relaxed'
+   * @param modelId - model id (e.g. 'flux1-schnell-fp8')
+   * @param forceRefresh - force refresh cache
+   * @returns {Promise<{
+   *   label: string;
+   *   id: string;
+   *   width: number;
+   *   height: number;
+   *   ratio: string;
+   *   aspect: string;
+   * }[]>}
+   */
+  async getSizePresets(network: SupernetType, modelId: string, forceRefresh = false) {
+    const key = `${network}-${modelId}`;
+    const cached = sizePresetCache.read(key);
+    if (cached && !forceRefresh) {
+      return cached;
+    }
+    const data = await this.client.socket.get<SizePreset[]>(
+      `/api/v1/size-presets/network/${network}/model/${modelId}`
+    );
+    sizePresetCache.write(key, data);
+    return data;
+  }
+
+  /**
+   * Get available models and their worker counts. Normally, you would get list once you connect
+   * to the server, but you can also call this method to get the list of available models manually.
+   * @param network
+   */
+  async getAvailableModels(network: SupernetType): Promise<AvailableModel[]> {
+    const workersByModelSid = await this.client.socket.get<Record<string, number>>(
+      `/api/v1/status/network/${network}/models`
+    );
+    const supportedModels = await this.getSupportedModels();
+    return Object.entries(workersByModelSid).map(([sid, workerCount]) => {
+      const SID = Number(sid);
+      const model = supportedModels.find((m) => m.SID === SID);
+      return {
+        id: model?.id || sid,
+        name: model?.name || sid.replace(/-/g, ' '),
+        workerCount
+      };
+    });
   }
 }